requirements.txt
pandas==1.5.2
plotly==5.11.0
praw==7.6.1
python-dotenv==0.21.0
streamlit==1.16.0
transformers==4.25.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu